refactored and started on multiple runs ltsm
